move utility cleancopy method
